move fonts to utils module
