+++ dcspy/__init__.py
@@ -1,10 +1,8 @@
 from logging import getLogger
-from os import name, path
+from os import name
 from platform import architecture, uname, python_implementation, python_version
-from sys import platform, prefix
+from sys import platform
 from typing import NamedTuple
-
-from PIL import ImageFont
 
 from dcspy.log import config_logger
 
@@ -21,11 +19,3 @@ config_logger(LOG)
 LOG.debug(f'Arch: {name} / {platform} / {" / ".join(architecture())}')
 LOG.debug(f'Python: {python_implementation()}-{python_version()}')
 LOG.debug(f'{uname()}')
-
-if platform == 'win32':
-    DEDFONT_11 = ImageFont.truetype(path.join(prefix, 'dcspy_data', 'falconded.ttf'), 11)
-    FONT_11 = ImageFont.truetype('consola.ttf', 11)
-    FONT_16 = ImageFont.truetype('consola.ttf', 16)
-else:
-    FONT_11 = ImageFont.truetype('DejaVuSansMono.ttf', 11)
-    FONT_16 = ImageFont.truetype('DejaVuSansMono.ttf', 16)
